+++ pimd/pim6_cmd.c
@@ -1992,6 +1992,18 @@ DEFPY (clear_ipv6_pim_oil,
 	return CMD_SUCCESS;
 }
 
+DEFPY (clear_ipv6_mroute_count,
+       clear_ipv6_mroute_count_cmd,
+       "clear ipv6 mroute [vrf NAME]$name count",
+       CLEAR_STR
+       IPV6_STR
+       MROUTE_STR
+       VRF_CMD_HELP_STR
+       "Route and packet count data\n")
+{
+	return clear_ip_mroute_count_command(vty, name);
+}
+
 void pim_cmd_init(void)
 {
 	if_cmd_init(pim_interface_config_write);
@@ -2102,4 +2114,5 @@ void pim_cmd_init(void)
 	install_element(ENABLE_NODE, &clear_ipv6_pim_statistics_cmd);
 	install_element(ENABLE_NODE, &clear_ipv6_mroute_cmd);
 	install_element(ENABLE_NODE, &clear_ipv6_pim_oil_cmd);
+	install_element(ENABLE_NODE, &clear_ipv6_mroute_count_cmd);
 }
