+++ pimd/pim_cmd_common.h
@@ -121,6 +121,7 @@ void show_mroute_count(struct pim_instance *pim, struct vty *vty,
 		       json_object *json);
 void show_mroute_summary(struct pim_instance *pim, struct vty *vty,
 			 json_object *json);
+int clear_ip_mroute_count_command(struct vty *vty, const char *name);
 struct vrf *pim_cmd_lookup(struct vty *vty, const char *name);
 void clear_mroute(struct pim_instance *pim);
 void clear_pim_statistics(struct pim_instance *pim);
